+++ src/Servers/Implementation.php
@@ -3,10 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Server\Control\Servers;
 
-use Innmind\Server\Control\Server\{
-    Processes,
-    Volumes,
-};
+use Innmind\Server\Control\Server\Processes;
 
 /**
  * @internal
@@ -15,7 +12,4 @@ interface Implementation
 {
     #[\NoDiscard]
     public function processes(): Processes;
-
-    #[\NoDiscard]
-    public function volumes(): Volumes;
 }