+++ src/Servers/Logger.php
@@ -3,10 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Server\Control\Servers;
 
-use Innmind\Server\Control\{
-    Server\Processes,
-    Server\Volumes,
-};
+use Innmind\Server\Control\Server\Processes;
 use Psr\Log\LoggerInterface;
 
 /**
@@ -15,7 +12,6 @@ use Psr\Log\LoggerInterface;
 final class Logger implements Implementation
 {
     private Processes $processes;
-    private Volumes $volumes;
 
     private function __construct(
         Implementation $server,
@@ -25,7 +21,6 @@ final class Logger implements Implementation
             $server->processes(),
             $logger,
         );
-        $this->volumes = new Volumes\Unix($this->processes);
     }
 
     public static function psr(Implementation $server, LoggerInterface $logger): self
@@ -37,11 +32,5 @@ final class Logger implements Implementation
     public function processes(): Processes
     {
         return $this->processes;
-    }
-
-    #[\Override]
-    public function volumes(): Volumes
-    {
-        return $this->volumes;
     }
 }